# services/payments-svc/internal/server/openapi.yaml
...
openapi: 3.0.3
info:
    title: Reserve interbank settlement payment hub API
    version: 1.0.0
paths:
...
    /v1/ops/defunds:
        post:
            tags:
                - BankOperationsService
            description: |-
                RequestDefund asks for settlement money back to the master account as
                 reserves (maker); the amount is earmarked at once.
            operationId: BankOperationsService_RequestDefund
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.RequestDefundRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Defund'
    /v1/ops/defunds/{id}/approve:
        post:
            tags:
                - BankOperationsService
            description: ApproveDefund is the checker; never the requester, always the same bank.
            operationId: BankOperationsService_ApproveDefund
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ApproveDefundRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Defund'
    /v1/ops/funding:
        post:
            tags:
                - BankOperationsService
            description: |-
                FundSettlement sends reserves into the reserve account and the operator
                 issues settlement money against them: Fedwire BTRC while Fedwire is open,
                 FedNow LMT1 while it is closed.
            operationId: BankOperationsService_FundSettlement
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.FundSettlementRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.FundSettlementResponse'
...
    /v1/ops/holds/{id}/block:
        post:
            tags:
                - BankOperationsService
            description: |-
                BlockHold: a sanctioned party has an interest in the funds. They move to
                 the bank's blocked account (BLCK) and a report is due to OFAC within 10
                 business days (31 CFR 501.603).
            operationId: BankOperationsService_BlockHold
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.HoldDecisionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/ops/holds/{id}/reject:
        post:
            tags:
                - BankOperationsService
            description: |-
                RejectHold: prohibited, but no sanctioned party has an interest. Nothing
                 moves (RR04); reported within 10 business days (31 CFR 501.604).
            operationId: BankOperationsService_RejectHold
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.HoldDecisionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/ops/holds/{id}/release:
        post:
            tags:
                - BankOperationsService
            description: ReleaseHold clears a false positive; the payment proceeds.
            operationId: BankOperationsService_ReleaseHold
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.HoldDecisionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/ops/intraday-draws:
        post:
            tags:
                - BankOperationsService
            description: |-
                DrawIntraday borrows settlement money from the intraday pool against
                 collateral at the pool's advance rate; payments waiting for liquidity
                 then proceed.
            operationId: BankOperationsService_DrawIntraday
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.DrawIntradayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.DrawIntradayResponse'
    /v1/ops/intraday-draws/{id}/repay:
        post:
            tags:
                - BankOperationsService
            description: RepayIntraday repays a draw with its interest; the collateral returns.
            operationId: BankOperationsService_RepayIntraday
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.RepayIntradayRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.IntradayDraw'
...
    /v1/ops/payments:
        get:
            tags:
                - BankOperationsService
            description: ListBankPayments lists the bank's clients' payments, with staff detail.
            operationId: BankOperationsService_ListBankPayments
            parameters:
                - name: status
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListBankPaymentsResponse'
    /v1/ops/payments/{id}:
        get:
            tags:
                - BankOperationsService
            description: GetBankPayment returns one of the bank's clients' payments, with staff detail.
            operationId: BankOperationsService_GetBankPayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
...
components:
    schemas:
...
        api.payments.v1.ApproveDefundRequest:
            type: object
            properties:
                id:
                    type: string
...
        api.payments.v1.Defund:
            type: object
            properties:
                id:
                    type: string
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                requestedBy:
                    type: string
                approvedBy:
                    type: string
                status:
                    type: string
                createdAt:
                    type: string
                    format: date-time
...
        api.payments.v1.DrawIntradayRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
        api.payments.v1.DrawIntradayResponse:
            type: object
            properties:
                draw:
                    $ref: '#/components/schemas/api.payments.v1.IntradayDraw'
                paymentsReleased:
                    type: integer
                    format: int32
...
        api.payments.v1.FundSettlementRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
        api.payments.v1.FundSettlementResponse:
            type: object
            properties:
                instrument:
                    type: string
                status:
                    type: string
                reason:
                    type: string
                paymentsReleased:
                    type: integer
                    format: int32
        api.payments.v1.HoldDecisionRequest:
            type: object
            properties:
                id:
                    type: string
                note:
                    type: string
...
        api.payments.v1.Payment:
            type: object
            properties:
                id:
                    type: string
                endToEndId:
                    type: string
                uetr:
                    type: string
                idempotencyKey:
                    type: string
                organizationId:
                    type: string
                debtorAccount:
                    type: string
                creditor:
                    $ref: '#/components/schemas/api.payments.v1.Party'
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                priority:
                    type: string
                remittance:
                    type: string
                status:
                    type: string
                    description: |-
                        AWAITING_APPROVAL | IN_PROCESS | ON_HOLD | AWAITING_LIQUIDITY |
                         QUEUED_FOR_NETTING | SETTLED | REJECTED | CANCELLED | EXPIRED | BLOCKED
                isoStatus:
                    type: string
                    description: RCVD | ACTC | PATC | ACCP | PDNG | ACSP | ACSC | ACCC | RJCT | CANC | BLCK
                reasonCode:
                    type: string
                route:
                    type: string
                    description: instant, 24x7 | netting cycle | book transfer, same bank
                routeNote:
                    type: string
                payeeCheck:
                    $ref: '#/components/schemas/api.payments.v1.PayeeCheck'
                approvalsRequired:
                    type: integer
                    format: int32
                approvals:
                    type: array
                    items:
                        type: string
                createdBy:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                settledAt:
                    type: string
                    format: date-time
                ofacReportDue:
                    type: string
                    format: date-time
                history:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.PaymentEvent'
                fileId:
                    type: string
                sanctionsMatch:
                    type: string
            description: Payment is a payment order and its lifecycle.
...
        api.payments.v1.RepayIntradayRequest:
            type: object
            properties:
                id:
                    type: string
        api.payments.v1.RequestDefundRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
...